// filink-web/branches/20190223/src/app/business-module/alarm/alarm-language.interface.ts
export interface AlarmLanguageInterface {
    id: string;
    serialNumber: string;
    trapOid: string;
    alarmName: string;
    alarmCode: string;
    alarmContent: string;
    alarmType: string;
    alarmSource: string;
    alarmSourceType: string;
    alarmSourceTypeid: string;
    area: string;
    areaName: string;
    orderId: string;
    address: string;
    alarmFixedLevel: string;
    alarmLevel: string;
    alarmDefaultLevel: string;
    customlevel: string;
    alarmAutomaticConfirmation: string;
    alarmobject: string;
    responsibleDepartment: string;
    alarmBeginTime: string;
    alarmNearTime: string;
    alarmHappenCount: string;
    alarmCleanStatus: string;
    alarmCleanTime: string;
    alarmCleanType: string;
    alarmContinousTime: string;
    alarmCleanPeopleId: string;
    alarmConfirmStatus: string;
    alarmConfirmTime: string;
    alarmConfirmPeopleId: string;
    alarmCleanPeopleNickname: string;
    alarmConfirmPeopleNickname: string;
    extraMsg: string;
    remark: string;
    deleteHandle: string;
    update: string;
    add: string;
    operate: string;
    location: string;
    view: string;
    buildOrder: string;
    alarmConfirm: string;
    alarmClean: string;
    historyAlarmSet: string;
    alarmFiltration: string;
    alarmLevelSet: string;
    urgent: string;
    main: string;
    secondary: string;
    prompt: string;
    alarmLevelCode: string;
    alarmLevelColor: string;
    alarmLevelSound: string;
    isPlay: string;
    playCount: string;
    urgentAlarm: string;
    mainAlarm: string;
    secondaryAlarm: string;
    promptAlarm: string;
    red: string;
    orange: string;
    yellow: string;
    blue: string;
    delayTime: string;
    yes: string;
    no: string;
    noClean: string;
    isClean: string;
    deviceClean: string;
    isConfirm: string;
    noConfirm: string;
    currentAlarmSetTips: string;
    alarmFilter: string;
    enable: string;
    disable: string;
    brightYellow: string;
    deepBlue: string;
    green: string;
    editAlertSettings: string;
    cancelText: string;
    okText: string;
    save: string;
    historyAlarmSettings: string;
    editAlarmLevel: string;
    alarm: string;
    enterNormalTime: string;
    confirmAgain: string;
    pleaseChoose: string;
    cleanAgain: string;
    opticalBox: string;
    well: string;
    distributionFrame: string;
    junctionBox: string;
    splittingBox: string;
    enterMaxTime: string;
    name: string;
    filterConditions: string;
    opertionUser: string;
    openStatus: string;
    creatTime: string;
    isInventory: string;
    startTime: string;
    endTime: string;
    isInventoryData: string;
    isNoStartUsing: string;
    type: string;
    status: string;
    isNoStored: string;
    particulars: string;
    delete: string;
    addAlarmFiltration: string;
    updateAlarmFiltration: string;
    month: string;
    day: string;
    hour: string;
    fundamental: string;
    filtration: string;
    pleaseCheckThe: string;
    alarmAffirm: string;
    year: string;
    alarmAffirmClear: string;
}